# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Goodfood 3.0 Users Microservice",
    "version": "1.0.0",
    "title": "Goodfood 3.0 Users Service API"
  },
...
  "paths": {
...
    "/user": {
      "post": {
        "tags": ["user"],
        "summary": "Create user",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid form submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "UpdateUserRequestBody": {
        "type": "object",
        "properties": {
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "Last"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        }
      },
      "CreateUserRequestBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "fsa": {
                "type": "string",
                "pattern": "/^[A-Z][0-9][A-Z]$/g",
                "example": "H3A"
              },
              "referrer_id": {
                "description": "A GF2.0 GiveGet referrer id. 'g' - refers to an email invite. 'l' - refers to a link invite.",
                "type": "string"
              }
            },
            "required": ["fsa", "email", "firebase_id", "language"]
          },
          {
            "$ref": "#/components/schemas/UpdateUserRequestBody"
          }
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2",
            "description": "firebase_id comes from the ID generated on Firebase Authentication when the user is registered there. It is a random string of 28 characters with a mix of upper- and lower-case characters and numbers. It must be unique."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "McLasty"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        },
        "required": ["id", "firebase_id", "email", "language"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "error description"
          },
          "meta": {
            "description": "Structured representation of the error",
            "example": {
              "errors": [["The fsa field is required."]]
            }
          },
          "status_code": {
            "type": "number"
          }
        }
      },
...
    },
...
  }
}